Fix sitemap with disabled pages (fix #206)
Recommendation in pelican < 3.4 was to disable pages using
a False boolean (http://docs.getpelican.com/en/3.3.0/settings.html#basic-settings)
It's now recommended to use '' and the sitemap plugins tries to
join paths using those variables, which won't work with a bool.
This checks that we have a truthy value before joining.